Who Am I? MVP
This adds a basic `whoami` subcommand to print out the username, GUID,
and email of the logged in CF CLI user.

We have to add explicit help functions for each subcommand because the
CF CLI will parse any -h or --help flags and not call the plugin.

There is still a default help command which incorrectly tells the user
to use -h or --help for subcommand instructions. This will be addressed
in a future commit.